// payload-types.ts
export interface ImageGalleryBlock {
  images?:
    | {
        /**
         * Cloudinary URL
         */
        imageUrl: string;
        imageAlt?: string | null;
        alignment?:
          | ('top-left' | 'top' | 'top-right' | 'left' | 'center' | 'right' | 'bottom-left' | 'bottom' | 'bottom-right')
          | null;
        id?: string | null;
      }[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'imageGallery';
}